pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "playwright-practice"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.58.0,<2.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "requests>=2.32.0,<3.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
